requirements.txt
# -----------------------------------------------------------------------
# Core runtime
# -----------------------------------------------------------------------
# NOTE: apache-airflow is NOT listed here — it is already in the base image.

# FAB auth manager — required by Airflow 3.x for user authentication
apache-airflow-providers-fab

# Standard operators (ShortCircuitOperator, PythonOperator, etc.)
apache-airflow-providers-standard

# PostgreSQL — main database
psycopg2-binary


# GitHub REST API client
PyGithub

# HTTP client
requests

# Load environment variables from .env file
python-dotenv

# Retry decorator
tenacity